face_reg2/requirements.txt
click==8.0.4
cmake==3.27.2
colorama==0.4.5
dlib==19.8.1
face-recognition==1.3.0
face-recognition-models==0.3.0
importlib-metadata==4.8.3
numpy==1.19.5
opencv-python==4.5.4.58
Pillow==8.4.0
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
typing_extensions==4.1.1
zipp==3.6.0
